// types.ts
export interface Program {
  id: string;
  title: string;
  description: string;
  duration: string;
  icon: string;
  color: string;
  link: string;
}

export interface Benefit {
  title: string;
  description: string;
  icon: string;
}

export interface PartnershipFeature {
  title: string;
  description: string;
  icon: string;
}